Refactor almost_linear and Bernoulli to classes
